refactor: closeModal util 함수로 분리
