+++ src/components/CustomModal/index.ts
@@ -1,4 +1,5 @@
 import './style.css';
+import { closeModal } from '../../utils';
 
 class CustomModal extends HTMLElement {
   constructor() {
@@ -54,24 +55,7 @@ class CustomModal extends HTMLElement {
     const modal = this?.shadowRoot?.querySelector('.modal');
     modal
       ?.querySelector('.modal-backdrop')
-      ?.addEventListener('click', this.#closeModal);
-  }
-
-  #closeModal() {
-    const modalEl = document
-      .querySelector('custom-modal')
-      ?.shadowRoot?.querySelector('.modal');
-
-    if (modalEl) {
-      modalEl.classList.toggle('open');
-      const childSlotEl = document.querySelector('[slot="child"]');
-      if (childSlotEl) {
-        childSlotEl.innerHTML = '';
-      }
-
-      const bodyEl = document.querySelector('body');
-      if (bodyEl) bodyEl.style.overflowY = 'scroll';
-    }
+      ?.addEventListener('click', closeModal);
   }
 }
