+++ src/utils/index.ts
@@ -1,0 +1,4 @@
+import closeModal from './closeModal';
+import findParentBox from './findParentBox';
+
+export { closeModal, findParentBox };
